refactor: move tee body to a separate file
